[extractor/redbee] Unify and update extractors (#4479)
Closes #4443
Authored by: elyse0
